Added OpenAPI spec + swagger utils

+++ scripts/start.sh
@@ -5,5 +5,10 @@ then
   curl -L https://github.com/vrana/adminer/releases/download/v4.8.1/adminer-4.8.1.php -o backend/docker/adminer/adminer-4.8.1.php
 fi
 
-docker compose down --remove-orphans --rmi local --volumes
+COMMAND=$1
+
+if [[ $COMMAND == 'clean' ]]; then
+  docker compose down --remove-orphans --rmi local --volumes
+fi
+
 docker compose --env-file backend/.env.docker up -d
